Controller breakout for blog and comments
